Refactor to move pool execution to PadatiousService
